+++ apps/kinflow_app/android/app/src/main/kotlin/me/newlines/kinflow/MainActivity.kt
@@ -1,0 +1,5 @@
+package me.newlines.kinflow
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()